Theoretically fixed
No errors are popping up, so hopefully this is fixed. Honestly, I am not sure what to do anymore with this, but I have to figure it out somehow. We shall see what happens. Also, on an unrelated note, I doubt anyone will ever bother reading a summary of changes that is many sentences long. Eheu! At least this helps me to get my thoughts in order and out on something. I suppose it is best just to dump everything here now. Ah well... Back to trying to get everything to be fixed.