Updating the Readme
Now I know the formatting for .md files, as well as coming up with the idea to use the readme of small projects as brain dumps, all dated and logged neatly. Hopefully I can keep this trend up in the future, but I'm not sure if I will be quite that dutiful. I don't know quite why I have been writing essays for the 'brief summaries' that are recommended for commits, but it is what I have been doing. Ah well. This should work.