Commenting out existing code
In preparation of the clean slate, I have completely commented out all of the code by Ctrl + A, followed with a swift Ctrl + /.

I still need help and to get sleep because I feel like I'm going to fall asleep any second here. I'm going to stand up, go to the restroom, and return. I need my blood flowing and some goddamn energetic music. This swing, blues, and western music is all well and good, but I'm falling asleep because it is just too damn relaxing.
also hello to anyone reading this for whatever reason. I hope this doesn't get you down too much, but it is what it is. If you wish to continue reading you may, but I do not think that is the best idea. Adieu!